Create a lints workspace to house lint rules
