+++ support/lints/error_enum/.gitignore
@@ -1,0 +1,1 @@
+/target
