+++ support/lints/error_enum/README.md
