+++ support/lints/error_enum/ui/main.stderr
